refactor(facturacion-exportacion): usar BigDecimal para cantidades decimales

+++ src/main/java/eterea/core/service/hexagonal/facturacion/arca/exportacion/application/service/FacturaElectronicaExportacionService.java
@@ -42,7 +42,7 @@ public class FacturaElectronicaExportacionService {
                 .map(movimiento -> new FacturaExportacionFacturadorItem(
                         movimiento.getArticuloId(),
                         movimiento.getDescripcion(),
-                        movimiento.getCantidad().abs().intValue(),
+                        movimiento.getCantidad().abs(),
                         7,
                         movimiento.getPrecioUnitario(),
                         movimiento.getTotal()))
